requirements.txt
numpy
scipy
pytest-cov
codecov
matplotlib
